User Id changed to Long to Integer

+++ src/main/java/com/capgemini/flightbookingsystem/entities/User.java
@@ -16,7 +16,7 @@ public class User {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "user_id")
-	private Long userId;
+	private Integer userId;
 
 	@NotBlank(message = "User name is mendentory.")
 	@Column(name = "name")
@@ -42,7 +42,7 @@ public class User {
 	public User() {
 	}
 
-	public User(Long userId, String name, String email, String password, String phoneNumber, String passportNumber) {
+	public User(Integer userId, String name, String email, String password, String phoneNumber, String passportNumber) {
 		super();
 		this.userId = userId;
 		this.name = name;
@@ -53,11 +53,11 @@ public class User {
 
 	}
 
-	public Long getUserId() {
+	public Integer getUserId() {
 		return userId;
 	}
 
-	public void setUserId(Long userId) {
+	public void setUserId(Integer userId) {
 		this.userId = userId;
 	}
